+++ pkg/worker/nopio.go
@@ -1,0 +1,7 @@
+package worker
+
+type nopWriter struct{}
+
+func (nopWriter) Write(bytes []byte) (int, error) {
+	return len(bytes), nil
+}